+++ test/getHeaders.spec.js
@@ -6,14 +6,25 @@ describe('getHeaders function', () => {
     localStorage.clear();
   });  
 
-  test('should return only Auth header if FormData params present', () => {
+  test('should ommit json headers if FormData params present', () => {
     localStorage.setItem('_token', 'MOCK-JWT-TOKEN');
-    expect(Api.getHeaders(new FormData)).toEqual({
-      'Authorization': 'Bearer MOCK-JWT-TOKEN',
-    });
+    const notToHave = {
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',      
+    };
+    expect(Api.getHeaders(new FormData))
+      .toEqual(expect.not.objectContaining(notToHave));
   });
 
-  test('should return correct headers with normal params', () => {
+  test('should have auth header if FormData params present', () => {
+    localStorage.setItem('_token', 'MOCK-JWT-TOKEN');
+    expect(Api.getHeaders(new FormData))
+      .toEqual({
+        'Authorization': 'Bearer MOCK-JWT-TOKEN',
+      });
+  });  
+
+  test('should return correct headers with params', () => {
     expect(Api.getHeaders({ foo: 'bar' })).toEqual({
       'Content-Type': 'application/json',
       'Accept': 'application/json',
@@ -21,7 +32,7 @@ describe('getHeaders function', () => {
     });
   });
 
-  test('should return correct headers with auth params', () => {
+  test('should return correct headers with auth', () => {
     localStorage.setItem('_token', 'MOCK-JWT-TOKEN');
     expect(Api.getHeaders({ foo: 'bar' })).toEqual({
       'Content-Type': 'application/json',
@@ -29,4 +40,32 @@ describe('getHeaders function', () => {
       'Authorization': 'Bearer MOCK-JWT-TOKEN',
     });
   });
+
+  test('should merge headers without params', () => {
+    expect(Api.getHeaders({}, { foobar: 'barfoo' })).toEqual({
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+      'Authorization': '',
+      'foobar': 'barfoo',
+    });
+  });  
+
+  test('should merge headers with params', () => {
+    expect(Api.getHeaders({ fickle: 'feline' }, { foobar: 'barfoo' })).toEqual({
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+      'Authorization': '',
+      'foobar': 'barfoo',
+    });
+  });    
+
+  test('should merge headers with params and auth', () => {
+    localStorage.setItem('_token', 'MOCK-JWT-TOKEN');
+    expect(Api.getHeaders({ fickle: 'feline' }, { foobar: 'barfoo' })).toEqual({
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+      'Authorization': 'Bearer MOCK-JWT-TOKEN',
+      'foobar': 'barfoo',
+    });
+  });      
 });